requirements.txt
python-binance
TA-Lib
numpy
flask
backtrader
matplotlib==3.2.2